src/requirements.txt
bson==0.5.10
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
image==1.5.32
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.1.2
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.1
six==1.12.0
sqlparse==0.3.1
Werkzeug==1.0.1
